Move test-only code into Cas1 Placement Req Test
This commit moves a transform function that was only used by tests from `main` into the corresponding tests
